[clean] Loadingを共通コンポーネントに変更
