+++ src/view/pages/AddHistory/index.tsx
@@ -1,16 +1,22 @@
 import { useState } from 'react'
 import { StoreWatchHistoryProgressType } from 'src/common/utils/db/watchHistory'
-import Ad from 'src/view/components/atoms/Ad'
 import ColumnContent from 'src/view/components/templates/ColumnContent'
-import styled from 'styled-components'
 import InitPage from '../../components/utils/InitPage'
 import Abstract from './Abstract'
 import CompleteButton from './CompleteButton'
 import Description from './Description'
-import Loading from './Loading'
 import UploadButton from './UploadButton'
+import Loading from 'src/view/components/organisms/Loading'
 
 export type ProgressType = StoreWatchHistoryProgressType | 'ready' | 'init'
+
+const progressSnippet: { [index in ProgressType]: string } = {
+  ready: '',
+  init: '初期化中',
+  parse: 'JSONデータを解析中',
+  store: 'データベースに格納中',
+  finished: '',
+}
 
 const AddHistory: React.FC = () => {
   const [progress, setProgress] = useState<ProgressType>('ready')
@@ -21,7 +27,7 @@ const AddHistory: React.FC = () => {
       {progress == 'ready' || progress == 'finished' ? (
         <Abstract />
       ) : (
-        <Loading progress={progress} />
+        <Loading loadingSnippet={progressSnippet[progress]} />
       )}
       {progress == 'ready' ? (
         <UploadButton progressSetter={setProgress} />
